% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.r
\name{download.smap.l3}
\alias{download.smap.l3}
\title{Download SMAP data for date}
\usage{
download.smap.l3(date, data.dir = "smap_ap", dataset.id = "SM_AP")
}
\description{
Checks whether download already exsts in data directory. If local copy does not exist, then performs FTP download.
}
\examples{
download.smap.l3("2015-09-11")
}
\keyword{download}

