% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_smap.R
\name{extract_smap}
\alias{extract_smap}
\title{Extracts contents of SMAP data files}
\usage{
extract_smap(file, group, dataset)
}
\arguments{
\item{file}{A character string path to a file on the local file system, e.g.,
produced by \code{download_smap()} that specifies input data files.}

\item{group}{The group in the HDF5 file containing data to extract.}

\item{dataset}{The dataset to extract.}
}
\value{
Returns a Raster object.
}
\description{
Extracts a datasets from a SMAP data file.
}
\details{
The arguments \code{group} and \code{dataset} must refer specifically  the
group and name within group for the input file, such as can be obtained with
\code{list_smap()}. This function will extract that particular dataset,
returning a Raster object.
}
\examples{
files <- find_smap(id = "SPL4SMGP", date = "2015.03.31")
downloads <- download_smap(files[1, ])
h5_data <- subset(downloads, file_ext == '.h5')
sm_raster <- extract_smap(h5_data$file[1],
                             group = 'Geophysical_Data',
                             dataset = 'leaf_area_index')
}

