% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smap.R
\docType{package}
\name{foo}
\alias{foo}
\alias{foo-package}
\title{smapr: A package for acquisition and processing of NASA SMAP data.}
\description{
The smapr package provides a means to programatically search for and acquire
NASA Soil Moisture Active Passive (SMAP) data.
}
\section{Foo functions}{

The foo functions ...
}

