% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.r
\name{global.smap.l3}
\alias{global.smap.l3}
\title{Download global SMAP data beginning at date}
\usage{
global.smap.l3(begin, bounding.box = NULL, ...)
}
\description{
It takes several days for SMAP to cover the entire globe. This function provides global
coverage beginning at given date.
}
\examples{
global.smap.l3("2015-09-11")
}
\keyword{download}

