% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Prints "Hello...<some name>.}
\usage{
hello(name)
}
\arguments{
\item{name}{A character string that defines a name, for instance of a person
or animal.}
}
\value{
A character string that is "Hello...<your name>".
}
\description{
This function has one input \code{name}, and prints a character string that
is "Hello...<your name>".
}
\details{
There aren't many details to include.
}
\examples{
hello(name = 'Clarice')
}

