% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.r
\name{read.smap.l3}
\alias{read.smap.l3}
\title{Read SMAP data for date}
\usage{
read.smap.l3(date, data.dir = "/home/homer/data/smap_ap",
  bounding.box = central.india.bbox)
}
\description{
Returns a dataframe withe SMAP data. Will download if necessary.
}
\examples{
download.smap.l3("2015-09-11")
}
\keyword{download}

