% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.r
\name{smap.filename}
\alias{smap.filename}
\title{SMAP filename}
\usage{
smap.filename(date, dataset.id = "SM_AP", gzip = TRUE)
}
\description{
Generates the filenames for HDF5. Incomplete.
}
\examples{
smap.filename("2015-09-11")
}
\keyword{download}

