% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.r
\name{smap.url}
\alias{smap.url}
\title{SMAP URL}
\usage{
smap.url(date, dataset.id = "SM_AP", gzip = TRUE)
}
\description{
Returns the FTP URL from which SMAP can be downloaded. By default fetches gzip compressed file.
}
\examples{
smap.url("2015-09-11")
}
\keyword{download}

