% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_smap.R
\name{download_smap}
\alias{download_smap}
\title{Download SMAP data}
\usage{
download_smap(files_to_download, directory = NULL)
}
\arguments{
\item{files_to_download}{A \code{data.frame} produced by \code{find_smap()} that
specifies data files to download.}

\item{directory}{A directory path in which to save data, specified as a character
string. If left as \code{NULL}, data are stored in a user's cache directory.}
}
\value{
Returns a \code{data.frame} that appends a column called \code{file}
to the input data frame, which consists of a character vector of file paths
to the downloaded files.
}
\description{
This function downloads SMAP data in hdf5 format.
}
\examples{
files <- find_smap(id = "SPL4SMGP", date = "2015.03.31")
# files[1, ] refers to the first available data file
download_smap(files[1, ])
}

