% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smapr-package.R
\docType{package}
\name{smapr-package}
\alias{smapr-package}
\title{smapr: A package for acquisition and processing of NASA SMAP data.}
\description{
The smapr package provides a means to programatically search for and acquire
NASA Soil Moisture Active Passive (SMAP) data.
}
\author{
Max Joseph \email{maxwell.b.joseph@colorado.edu}
}
\keyword{package}
